requirements.txt
pre-commit==4.0.0
flake8==6.1.0
pylint==3.0.0
python-dotenv==1.0.1

amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.1
djangorestframework==3.15.2
docopt==0.6.2
executing==2.1.0
export==0.2.1
fastjsonschema==2.20.0
fonttools==4.54.1
idna==3.8
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
kombu==5.4.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
music==1.0.0b5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.1
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pyzmq==26.2.0
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
scipy==1.14.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.3
termcolor==2.4.0
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
